requirements.txt
Flask==2.1.3
python-dotenv==0.19.2
requests==2.26.0
PyPDF2==2.10.5
boto3==1.18.65
langchain==0.2.5
langchain_community==0.2.5
openai==1.34.0
tiktoken==0.7.0
werkzeug==2.0.3
faiss-gpu==1.7.2